move env to new file to prevent circular dependency
